cattle/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==4.2.9
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==10.2.0
pytz==2024.1
soupsieve==2.5
sqlparse==0.4.4
